pcf8523: cleanup
